refactor(sidebar-nav): slim API, docs, and demo data
