+++ src/app/(registry)/registry/[name]/page.tsx
@@ -151,7 +151,7 @@ export default function RegistryItemPage() {
 
   return (
     <div className="bg-surface-base min-h-screen w-full p-5 md:p-10">
-      <div className="mx-auto w-full max-w-[800px]">
+      <div className="mx-auto w-full max-w-[960px]">
         <ComponentPageLayout
           component={component}
           apiData={apiData || undefined}
